--- a/createMeasurements.py
+++ b/createMeasurements.py
@@ -2,7 +2,7 @@ import argparse
 import time
 
 import numpy as np
-import pandas as pd
+import polars as pl
 
 
 class CreateMeasurement:
@@ -432,11 +432,11 @@ class CreateMeasurement:
             self,
             std_dev: float = 10,
             records: int = 10_000_000
-    ) -> pd.DataFrame:
+    ) -> pl.DataFrame:
         ii = self.rng.integers(len(self.station_names), size=records)
         names = self.station_names[ii]
         temperatures = self.rng.normal(self.station_means[ii], std_dev)
-        return pd.DataFrame({"stations": names, "measurements": temperatures})
+        return pl.DataFrame({"stations": names, "measurements": temperatures})
 
     def generate_measurement_file(
             self,
@@ -456,7 +456,7 @@ class CreateMeasurement:
             for i in range(batches):
                 from_, to = batch_ends[i], batch_ends[i + 1]
                 data = self.generate_batch(std_dev, to - from_)
-                data.to_csv(f, sep=sep, float_format=".1", header=False, index=False)
+                data.write_csv(f, separator=sep, float_precision=1, include_header=False)
                 print(
                     f" - Wrote {to} measurements in {time.time() - start:.2f} seconds"
                 )