--- a/requirements.txt
+++ b/requirements.txt
@@ -1,3 +1,3 @@
-numpy==1.24.2
-polars==0.20.16
-tqdm
+numpy>=1.24.2
+polars>=0.20.0
+tqdm>=4.66.0