--- a/requirements.txt
+++ b/requirements.txt
@@ -1,2 +1,3 @@
 numpy==1.24.2
 polars==0.20.16
+tqdm